heroes-chatbot-final/backend/requirements.txt
fastapi
uvicorn
python-socketio[asyncio]
httpx>=0.27
openai
python-dotenv
pandas
openpyxl
beautifulsoup4>=4.12
lxml>=5.0
tldextract>=5.1
# optional (recommended for richer link extraction)
trafilatura>=1.7
pdfminer.six
pdfplumber
redis>=5.0